+++ Source/ViewControllerPresentationSpy/AlertVerifier.swift
@@ -102,6 +102,51 @@ public class AlertVerifier: NSObject {
     }
 }
 
+extension AlertVerifier {
+    public func verify(
+            title: String?,
+            message: String?,
+            animated: Bool,
+            preferredStyle: UIAlertController.Style = .alert,
+            presentingViewController: UIViewController? = nil,
+            file: StaticString = #file,
+            line: UInt = #line
+    ) {
+        if presentedCount == 0 {
+            XCTFail("present not called", file: file, line: line)
+            return
+        }
+        if presentedCount > 1 {
+            XCTFail("present called \(presentedCount) times", file: file, line: line)
+        }
+        XCTAssertEqual(self.title, title, "title")
+        XCTAssertEqual(self.message, message, "message")
+        if animated != self.animated {
+            if animated {
+                XCTFail("Expected animated present, but was not animated", file: file, line: line)
+            } else {
+                XCTFail("Expected non-animated present, but was animated", file: file, line: line)
+            }
+        }
+        if preferredStyle != self.preferredStyle {
+            switch preferredStyle {
+            case .actionSheet:
+                XCTFail("Expected preferred style .actionSheet, but was .alert", file: file, line: line)
+            case .alert:
+                XCTFail("Expected preferred style .alert, but was .actionSheet", file: file, line: line)
+            @unknown default:
+                print("preferred style not recognized, please contact maintainer")
+            }
+        }
+        if let presentingViewController = presentingViewController {
+            XCTAssertTrue(presentingViewController === self.presentingViewController,
+                    "Expected presenting view controller to be \(presentingViewController)), " +
+                            "but was \(String(describing: self.presentingViewController))",
+                    file: file, line: line)
+        }
+    }
+}
+
 @objc enum AlertVerifierErrors: Int, Error {
     case buttonNotFound
 }